main/java - answer,bookstory,question,reply

+++ sblib/src/main/java/org/sb/bookstory/domain/BookStoryVO.java
@@ -9,8 +9,8 @@ import lombok.ToString;
 @ToString
 public class BookStoryVO {
 
-	private long story_no;
+	private Long story_no;
 	private String story_title, story_content, story_author;
 	private Date story_regdate, story_updatedate;
-	private int story_readcount;
+	private Long story_readcount;
 }
